Implemented performance improvements for CSS-selection block detection and fix for connections remaining open.
There were a few naive mistakes in the original implementation, as well as a preference for not significantly modifying the existing AdBlock code.  Instead of calling "querySelectorAll", which returns a list of all DOM elements that match the given selector string, we now use "querySelector" as we really only care if at least one DOM element matches.  In addition, as we are no longer modifying the CSS, there is no need for an arbitrary delay to wait until ".sheet" exists.

Finally, I'd observed that some open network connections (requests for ad content, images etc.) were being held in an open state and never terminating when the page was redirected.  This was just bad manners.  Calling the "stop" method on the root-window prior to redirecting should hopefully fix this, but requires further testing.